+++ c++/pod5_format/c_api.cpp
@@ -438,7 +438,7 @@ pod5_error_t pod5_get_read_batch_row_info_data(
     }
 
     static_assert(
-        READ_BATCH_ROW_INFO_VERSION == READ_BATCH_ROW_INFO_VERSION_5,
+        READ_BATCH_ROW_INFO_VERSION == READ_BATCH_ROW_INFO_VERSION_6,
         "New versions must be explicitly loaded");
 
     auto load_common_v3_v4_fields = [](pod5::ReadTableRecordColumns const & cols,
@@ -456,7 +456,10 @@ pod5_error_t pod5_get_read_batch_row_info_data(
         typed_row_data->read_number = cols.read_number->Value(row);
         typed_row_data->start_sample = cols.start_sample->Value(row);
         typed_row_data->median_before = cols.median_before->Value(row);
-        typed_row_data->channel = cols.channel->Value(row);
+
+        // In V6 this will be overwritten with 32-bit data.
+        typed_row_data->channel =
+            cols.channel_32bit ? cols.channel_32bit->Value(row) : cols.channel_16bit->Value(row);
         typed_row_data->well = cols.well->Value(row);
         auto const & pore_type_col = cols.pore_type->indices();
         typed_row_data->pore_type =
@@ -482,17 +485,31 @@ pod5_error_t pod5_get_read_batch_row_info_data(
         return POD5_OK;
     };
 
-    if (struct_version == READ_BATCH_ROW_INFO_VERSION_3) {
+    auto load_common_v5_v6_fields =
+        [&load_common_v3_v4_fields](
+            pod5::ReadTableRecordColumns const & cols, std::size_t row, auto * typed_row_data) {
+            auto result = load_common_v3_v4_fields(cols, row, typed_row_data);
+            if (result != POD5_OK) {
+                return result;
+            }
+
+            typed_row_data->open_pore_level = cols.open_pore_level->Value(row);
+            typed_row_data->expected_open_pore_level = cols.expected_open_pore_level->Value(row);
+            typed_row_data->selected_read_level = cols.selected_read_level->Value(row);
+
+            return POD5_OK;
+        };
+
+    switch (struct_version) {
+    case READ_BATCH_ROW_INFO_VERSION_3: {
         auto typed_row_data = static_cast<ReadBatchRowInfoV3 *>(row_data);
 
         POD5_C_ASSIGN_OR_RAISE(auto cols, batch->batch.columns());
         *read_table_version = cols.table_version.as_int();
 
-        auto result = load_common_v3_v4_fields(cols, row, typed_row_data);
-        if (result != POD5_OK) {
-            return result;
-        }
-    } else if (struct_version == READ_BATCH_ROW_INFO_VERSION_4) {
+        return load_common_v3_v4_fields(cols, row, typed_row_data);
+    }
+    case READ_BATCH_ROW_INFO_VERSION_4: {
         auto typed_row_data = static_cast<ReadBatchRowInfoV4 *>(row_data);
 
         POD5_C_ASSIGN_OR_RAISE(auto cols, batch->batch.columns());
@@ -505,25 +522,38 @@ pod5_error_t pod5_get_read_batch_row_info_data(
 
         // This is the only difference between v3 and v4.
         typed_row_data->open_pore_level = cols.open_pore_level->Value(row);
-    } else if (struct_version == READ_BATCH_ROW_INFO_VERSION_5) {
+        break;
+    }
+    case READ_BATCH_ROW_INFO_VERSION_5: {
         auto typed_row_data = static_cast<ReadBatchRowInfoV5 *>(row_data);
 
         POD5_C_ASSIGN_OR_RAISE(auto cols, batch->batch.columns());
         *read_table_version = cols.table_version.as_int();
 
-        auto result = load_common_v3_v4_fields(cols, row, typed_row_data);
-        if (result != POD5_OK) {
+        if (auto result = load_common_v5_v6_fields(cols, row, typed_row_data); result != POD5_OK) {
             return result;
         }
+        break;
+    }
+    case READ_BATCH_ROW_INFO_VERSION_6: {
+        auto typed_row_data = static_cast<ReadBatchRowInfoV6 *>(row_data);
 
-        typed_row_data->open_pore_level = cols.open_pore_level->Value(row);
-        typed_row_data->expected_open_pore_level = cols.expected_open_pore_level->Value(row);
-        typed_row_data->selected_read_level = cols.selected_read_level->Value(row);
-    } else {
+        POD5_C_ASSIGN_OR_RAISE(auto cols, batch->batch.columns());
+        *read_table_version = cols.table_version.as_int();
+
+        if (auto result = load_common_v5_v6_fields(cols, row, typed_row_data); result != POD5_OK) {
+            return result;
+        }
+        typed_row_data->channel = cols.channel_32bit   ? cols.channel_32bit->Value(row)
+                                  : cols.channel_16bit ? cols.channel_16bit->Value(row)
+                                                       : 0;
+        break;
+    }
+    default:
         pod5_set_error(
             arrow::Status::Invalid("Invalid struct version '", struct_version, "' passed"));
         return g_pod5_error_no;
-    }
+    }  // End of switch(struct_version)
 
     return POD5_OK;
 }
@@ -1111,7 +1141,7 @@ pod5_error_t pod5_add_run_info(
 static bool check_read_data_struct(std::uint16_t struct_version, void const * row_data)
 {
     static_assert(
-        READ_BATCH_ROW_INFO_VERSION == READ_BATCH_ROW_INFO_VERSION_5,
+        READ_BATCH_ROW_INFO_VERSION == READ_BATCH_ROW_INFO_VERSION_6,
         "New versions must be explicitly loaded");
 
     if (!check_not_null(row_data)) {
@@ -1123,7 +1153,7 @@ static bool check_read_data_struct(std::uint16_t struct_version, void const * ro
         return false;
     }
 
-    auto check_common_v3_v4_fields = [](auto typed_row_data) -> bool {
+    auto const check_common_v3_v4_fields = [](auto typed_row_data) -> bool {
         return check_not_null(typed_row_data->read_id)
                && check_not_null(typed_row_data->read_number)
                && check_not_null(typed_row_data->start_sample)
@@ -1144,37 +1174,36 @@ static bool check_read_data_struct(std::uint16_t struct_version, void const * ro
                && check_not_null(typed_row_data->time_since_mux_change);
     };
 
-    if (struct_version == READ_BATCH_ROW_INFO_VERSION_3) {
+    auto const check_v5_v6_fields = [&](auto typed_row_data) -> bool {
+        return check_common_v3_v4_fields(typed_row_data)
+               && check_not_null(typed_row_data->open_pore_level)
+               && check_not_null(typed_row_data->expected_open_pore_level)
+               && check_not_null(typed_row_data->selected_read_level);
+    };
+
+    switch (struct_version) {
+    case READ_BATCH_ROW_INFO_VERSION_3: {
         auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV3 const *>(row_data);
-
-        if (!check_common_v3_v4_fields(typed_row_data)) {
-            return false;
-        }
+        return check_common_v3_v4_fields(typed_row_data);
     }
-
-    if (struct_version == READ_BATCH_ROW_INFO_VERSION_4) {
+    case READ_BATCH_ROW_INFO_VERSION_4: {
         auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV4 const *>(row_data);
-
-        if (!check_common_v3_v4_fields(typed_row_data)
-            || !check_not_null(typed_row_data->open_pore_level))
-        {
-            return false;
-        }
+        return check_common_v3_v4_fields(typed_row_data)
+               && check_not_null(typed_row_data->open_pore_level);
     }
-
-    if (struct_version == READ_BATCH_ROW_INFO_VERSION_5) {
+    case READ_BATCH_ROW_INFO_VERSION_5: {
         auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV5 const *>(row_data);
-
-        if (!check_common_v3_v4_fields(typed_row_data)
-            || !check_not_null(typed_row_data->open_pore_level)
-            || !check_not_null(typed_row_data->expected_open_pore_level)
-            || !check_not_null(typed_row_data->selected_read_level))
-        {
-            return false;
-        }
+        return check_v5_v6_fields(typed_row_data);
     }
-
-    return true;
+    case READ_BATCH_ROW_INFO_VERSION_6: {
+        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV6 const *>(row_data);
+        return check_v5_v6_fields(typed_row_data);
+    }
+    default:
+        assert(false && "Unknown READ_BATCH_ROW_INFO_VERSION!");
+        pod5_set_error(arrow::Status::Invalid("Unknown API version"));
+        return false;
+    }
 }
 
 static bool load_struct_row_into_read_data(
@@ -1185,13 +1214,13 @@ static bool load_struct_row_into_read_data(
     std::uint32_t row_id)
 {
     static_assert(
-        READ_BATCH_ROW_INFO_VERSION == READ_BATCH_ROW_INFO_VERSION_5,
+        READ_BATCH_ROW_INFO_VERSION == READ_BATCH_ROW_INFO_VERSION_6,
         "New versions must be explicitly loaded");
 
-    auto load_common_v3_v4_fields = [](std::unique_ptr<pod5::FileWriter> const & writer,
-                                       auto const * typed_row_data,
-                                       std::uint32_t row_id,
-                                       pod5::ReadData & read_data) {
+    auto const load_common_v3_v4_fields = [](std::unique_ptr<pod5::FileWriter> const & writer,
+                                             auto const * typed_row_data,
+                                             std::uint32_t row_id,
+                                             pod5::ReadData & read_data) {
         pod5::Uuid read_id_uuid{typed_row_data->read_id[row_id]};
 
         std::optional<pod5::ReadEndReason> end_reason_internal;
@@ -1271,34 +1300,58 @@ static bool load_struct_row_into_read_data(
         return true;
     };
 
-    // Version 0-2 are no longer supported for writing.
-    if (struct_version == READ_BATCH_ROW_INFO_VERSION_5) {
-        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV5 const *>(row_data);
-
+    auto const load_v5_v6_fields = [&load_common_v3_v4_fields](
+                                       std::unique_ptr<pod5::FileWriter> const & writer,
+                                       auto const * typed_row_data,
+                                       std::uint32_t row_id,
+                                       pod5::ReadData & read_data) -> bool {
         if (!load_common_v3_v4_fields(writer, typed_row_data, row_id, read_data)) {
             return false;
         }
         read_data.open_pore_level = typed_row_data->open_pore_level[row_id];
         read_data.expected_open_pore_level = typed_row_data->expected_open_pore_level[row_id];
         read_data.selected_read_level = typed_row_data->selected_read_level[row_id];
-    } else if (struct_version == READ_BATCH_ROW_INFO_VERSION_4) {
-        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV4 const *>(row_data);
+        return true;
+    };
 
+    switch (struct_version) {
+    case READ_BATCH_ROW_INFO_VERSION_3: {
+        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV3 const *>(row_data);
+        if (!load_common_v3_v4_fields(writer, typed_row_data, row_id, read_data)) {
+            return false;
+        }
+        break;
+    }
+    case READ_BATCH_ROW_INFO_VERSION_4: {
+        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV4 const *>(row_data);
         if (!load_common_v3_v4_fields(writer, typed_row_data, row_id, read_data)) {
             return false;
         }
         read_data.open_pore_level = typed_row_data->open_pore_level[row_id];
-    } else if (struct_version == READ_BATCH_ROW_INFO_VERSION_3) {
-        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV3 const *>(row_data);
+        break;
+    }
+    case READ_BATCH_ROW_INFO_VERSION_5: {
+        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV5 const *>(row_data);
 
-        if (!load_common_v3_v4_fields(writer, typed_row_data, row_id, read_data)) {
+        if (!load_v5_v6_fields(writer, typed_row_data, row_id, read_data)) {
             return false;
         }
-    } else {
+        break;
+    }
+    case READ_BATCH_ROW_INFO_VERSION_6: {
+        auto const * typed_row_data = static_cast<ReadBatchRowInfoArrayV6 const *>(row_data);
+
+        if (!load_v5_v6_fields(writer, typed_row_data, row_id, read_data)) {
+            return false;
+        }
+        break;
+    }
+    default:
+        // Version 0-2 are no longer supported for writing.
         pod5_set_error(
             arrow::Status::Invalid("Invalid writer struct version '", struct_version, "' passed"));
         return false;
-    }
+    }  // end of switch
     return true;
 };