+++ c++/pod5_format/migration/migration.h
@@ -83,6 +83,9 @@ POD5_FORMAT_EXPORT arrow::Result<MigrationResult> migrate_v3_to_v4(
 POD5_FORMAT_EXPORT arrow::Result<MigrationResult> migrate_v4_to_v5(
     MigrationResult && v4_input,
     arrow::MemoryPool * pool);
+POD5_FORMAT_EXPORT arrow::Result<MigrationResult> migrate_v5_to_v6(
+    MigrationResult && v5_input,
+    arrow::MemoryPool * pool);
 
 /*
 Perform physical file version migration to the minimum required on-disk version (schema)
@@ -122,8 +125,11 @@ inline arrow::Result<MigrationResult> migrate_to_latest(
     MigrationResult && input,
     arrow::MemoryPool * pool)
 {
-    ARROW_ASSIGN_OR_RAISE(auto result, migrate_v3_to_v4(std::move(input), pool));
-    return migrate_v4_to_v5(std::move(result), pool);
+    // Migrate functions must internally "pass-through" later versions rather than attempt to
+    // convert them.
+    ARROW_ASSIGN_OR_RAISE(auto v4_result, migrate_v3_to_v4(std::move(input), pool));
+    ARROW_ASSIGN_OR_RAISE(auto v5_result, migrate_v4_to_v5(std::move(v4_result), pool));
+    return migrate_v5_to_v6(std::move(v5_result), pool);
 }
 
 /*