+++ c++/pod5_format/version_constants.h
@@ -16,7 +16,8 @@ inline Version const kPod5VersionReadTableV3MigrationThreshold{0, 0, 38};
 
 inline Version const kPod5VersionReadTableV4{0, 3, 30};
 inline Version const kPod5VersionReadTableV5{0, 3, 40};
+inline Version const kPod5VersionReadTableV6{0, 3, 46};
 
-inline Version const kPod5VersionReadTableLatest = kPod5VersionReadTableV5;
+inline Version const kPod5VersionReadTableLatest = kPod5VersionReadTableV6;
 
 }  // namespace pod5