+++ python/pod5/src/tests/test_reader.py
@@ -130,13 +130,13 @@ class TestPod5Reader:
         with p5.Reader(POD5_PATH) as reader:
             assert isinstance(reader.path, Path)
             assert reader.path == POD5_PATH
-            assert reader.reads_table_version == 5
+            assert reader.reads_table_version == 6
             # Physical test file is v4 - which doesn't have v5 columns.
             # These columns are created virtually to avoid costly physical
             # file migration
             assert reader.original_file_version < packaging.version.Version("0.3.40")
             assert reader.physical_read_table_version == 4
-            assert reader.logical_read_table_version == 5
+            assert reader.logical_read_table_version == 6
             assert reader.reads_table_version == reader.logical_read_table_version
             assert "expected_open_pore_level" not in reader.read_table.schema.names
             assert "selected_read_level" not in reader.read_table.schema.names
@@ -175,12 +175,12 @@ class TestPod5Reader:
 
             assert isinstance(reader.get_batch(0), ReadRecordBatch)
 
-    def test_v3_read_table_v4_and_v5_fields_are_virtual(self) -> None:
+    def test_v3_read_table_later_version_fields_are_virtual(self) -> None:
         with p5.Reader(POD5_V3_PATH) as reader:
-            assert reader.reads_table_version == 5
+            assert reader.reads_table_version == 6
             assert reader.original_file_version < packaging.version.Version("0.3.30")
             assert reader.physical_read_table_version == 3
-            assert reader.logical_read_table_version == 5
+            assert reader.logical_read_table_version == 6
             assert "open_pore_level" not in reader.read_table.schema.names
             assert "expected_open_pore_level" not in reader.read_table.schema.names
             assert "selected_read_level" not in reader.read_table.schema.names
@@ -189,15 +189,17 @@ class TestPod5Reader:
             assert numpy.isnan(first_read.open_pore_level)
             assert numpy.isnan(first_read.expected_open_pore_level)
             assert numpy.isnan(first_read.selected_read_level)
+            # If the 1-biased channel field can't be read then 0 will be used as a substitute.
+            assert first_read.pore.channel != 0
 
     def test_v2_file_version_progression(self) -> None:
         with p5.Reader(POD5_V2_PATH) as reader:
             assert reader.original_file_version < packaging.version.Version("0.0.38")
             assert reader.physical_read_table_version == 3
-            assert reader.logical_read_table_version == 5
+            assert reader.logical_read_table_version == 6
 
     @pytest.mark.parametrize("random_read", [1], indirect=True)
-    def test_v5_read_table_fields_are_physical(
+    def test_latest_version_read_table_fields_are_physical(
         self, tmp_path: Path, random_read: p5.Read
     ) -> None:
         expected_open_pore_level = 2345.0
@@ -205,11 +207,14 @@ class TestPod5Reader:
         random_read.expected_open_pore_level = expected_open_pore_level
         random_read.selected_read_level = selected_read_level
 
-        path = tmp_path / "physical_v5.pod5"
+        # Create the file.
+        path = tmp_path / "physical_v6.pod5"
         with p5.Writer(path) as writer:
             writer.add_read(random_read)
 
+        # Now read it back.
         with p5.Reader(path) as reader:
+            assert reader.reads_table_version == 6
             assert "expected_open_pore_level" in reader.read_table.schema.names
             assert "selected_read_level" in reader.read_table.schema.names